Refactor temporary directory handling in tests
